Updated units and humidity & added docs
- The units parameter is now of enum type instead of string for nicer code. However, this way we can't handle incorrect values easily.

- Added the API documentation. (Experimental).

- Converted the humidity to int32 (%).

+++ WeatherService/Models/ResponseModel.cs
@@ -20,7 +20,7 @@ namespace WeatherService.Models
         {
             public long Date;
             public float? Temp;
-            public float? Humidity;
+            public int? Humidity;
             public string WeatherType;
             public string WeatherDescription;
             public int? Cloudiness;
@@ -35,7 +35,7 @@ namespace WeatherService.Models
             public long Date;
             public float? TempMin;
             public float? TempMax;
-            public float? Humidity;
+            public int? Humidity;
             public string WeatherType;
             public string WeatherDescription;
             public int? Cloudiness;
